Revise board, focussing on building into devices
New board layout has very few onboard components, and focuses on wiring
an IR LED into an existing device (next to its sensor), while leeching
power of the host device and piggybacking off its IR sensor.

There are jumpers for selecting input voltage (5V or 3.3V) which
determines where power gets routed to on the esp32 dev board, as well as
a bypassable voltage divider for the IR input. So if the host device
runs the IR sensor at 5V, it is selectable on this board, so it's
lowered to ~3.3V before going into the ESP32.

The IR Led connection to resistor/esp32 is also optional with a pair of
jumpers, so the connector can be used for other purposes if needed.

Lastly, there's a "protoboard" area for a small bit of custom circuitry
/ connectors, along with several power connection points.